now using Cookies instead of parameters where useful
When I say _where useful_, I mean wherever a parameter is passed that is used by the backend only and does not depend on the user. The identifier of a quotation is not in this pattern, but the parameter to check wether the last request was erroneous or not is.